+++ packages/viewer/src/index.html
@@ -35,8 +35,8 @@
       async function main() {
         const viewer = document.querySelector('#viewer');
 
-        viewer.token = 'MGVXVo6pigYTLpxDdBu-Z8237QO1Vot2SHpUsHCFoq4.u2cu1rQk29nKgt8F6KjGx-PHZ7ig4A4PunpvTd20G_4';
-        await viewer.load('urn:vertexvis:scene:51ba5e31-82c6-4f64-bcf5-349ec5802425');
+        viewer.token = 'gZ51MiERII5b1_Pd0dRSjn6xz2GQAI_8-zYhN0y0MQE.hfZ60f7r_xXbea3N3ZdvJwQGZcNNnpXf6W4_GHbYNzE';
+        await viewer.load('urn:vertexvis:scene:b9e39e5e-8e27-4d97-b1a3-ba5ada11ada5');
 
         viewer.addEventListener('tap', async function(event) {
           const { position } = event.detail;
@@ -44,7 +44,7 @@
           const raycaster = await scene.raycaster();
           scene
             .itemOperation()
-            .where(q => q.withItemId('02176fa0-1f1e-4b79-b1c5-bf84903bd6a1'))
+            .where(q => q.withItemId('a885f431-d39f-4293-b9d1-cb97a8573840'))
             .materialOverrideFromHex('#c9919f')
             .show()
             .execute();
